Updated settings pages
