+++ react-app/package.json
@@ -8,7 +8,6 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.11.62",
-    "@types/react-dom": "^18.0.6",
     "firebase": "^9.10.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -42,6 +41,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react": "^18.0.21"
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6"
   }
 }
