+++ react-app/src/settings-page/AdminSettingsPage.css
@@ -6,21 +6,30 @@
     padding: 0px;
 }
 .container{
-    display: flex;
-    flex-direction: column;
-
+  display: flex;
+  flex-flow: column;
 }
+
 .header{
-    height: 20vh;
+    height: 14vh;
     background-color: #FFFFFF;
-    display: flex;
-    justify-content: flex-start;
 }
 
 .body{
-    height: 80vh;
+    height: 10vh;
     background-color: #D7ECD3;
-}
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    flex: 1 1 100%;
+    overflow: auto;
+ }
+
+ .green-background{
+    height: 86vh;
+    background-color: #D7ECD3; 
+ }
 
 .rise-dc-img{
     position: absolute;
@@ -42,26 +51,32 @@
 .form{
     display: flex;
     flex-direction: column;
-    position: absolute;
+    position: relative;
     background-color: #FFFFFF;
-    top: 40%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    height: 300px;
-    width: 600px;
+    margin-top: 25px;
+    height: 310px;
+    width: 620px;
+    max-height: 100%;
+    max-width: 100%;
+    overflow-y: scroll;
+    flex-grow: 1;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 13px 27px -5px, rgba(0, 0, 0, 0.3) 0px 8px 16px -8px;
 }
 
 .form-second{
     display: flex;
     flex-direction: column;
-    position: absolute;
+    position:relative;
     background-color: #FFFFFF;
-    margin-top: 10px;
-    top: 78%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    height: 300px;
-    width: 600px; 
+    margin-top: 20px;
+    margin-bottom: 20px;
+    height: 310px;
+    width: 620px; 
+    max-height: 100%;
+    max-width: 100%;
+    overflow-y: scroll;
+    flex-grow: 1;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 13px 27px -5px, rgba(0, 0, 0, 0.3) 0px 8px 16px -8px;
 }
 
 .profile{
