+++ react-app/src/settings-page/ResetButton.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import "./AdminSettingsPage.css"
+
+export default function ResetButton() {
+    return (
+        <button className = "reset-button">Reset</button>
+        );
+    }
